R2.8.4 (#225)
* Adde BasicAll result processor

* Added ManagedProperties result processor. Fixed an issue with preset files containing select properties

* Retry if we get HTTP 401 after 30 minutes. Parameter to specify sleep time between query batches

* Checkpoint

* added support for multi geo

* Just documenting workaround for TLS issue.

* Enabled multigeo search and added a CrawledProperties result processor

* checkpoint

* checkpoint

* checkpoint

* cp

* Checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* cp

* Checkpoint

* Now reads environment variable PSSQT_DefaultAuthenticationMethod to determine which authentication method to use by default

* Made SetRequestAutheticationType protected virtual so that deriving classes can do "their own thing"

* Add some release notes

* Fixed broken multi-geo POST query

* Added a comment to highlight the need to format MultiGeoSearchConfiguration according to method type

* Yet another multi-geo post bug

* checkpoint

* added sponlinecredentials auth method

* Updated the authenticationmethod to allow credentials with SPOManagement.

* Rewrote the authentication logic. Command line always wins, if not use preset if applicable or if not guess based on url

* added some more verbose printing

* First update to 2.8.4.
Implemented tab completion on presets

* Added include personal OneDRive results

* Improved Error Reporting

* Try to fix xslt to work across browsers

* Refactored result processors into separate files in a sub folder

* Fixed typo in release notes
