+++ Solutions/SharePoint.Search.QueryTool/PSSQT/AuthenticationMethod.cs
@@ -4,14 +4,15 @@ namespace PSSQT
 {
     public enum PSAuthenticationMethod
     {
+        CurrentUser,
         Windows,
         SPO,
-        SPOManagement
-    }
+        SPOManagement   
+     }
 
     public class PSAuthenticationMethodFactory
     {
-        public static PSAuthenticationMethod DefaultAutenticationMethod()
+        public static PSAuthenticationMethod? DefaultAutenticationMethod()
         {
             string defaultAuthMethod = Environment.GetEnvironmentVariable("PSSQT_DefaultAuthenticationMethod");
 
@@ -23,7 +24,7 @@ namespace PSSQT
             }
 
 
-            return PSAuthenticationMethod.Windows;
+            return null;
         }
     }
 }
