+++ Solutions/SharePoint.Search.QueryTool/PSSQT/Helpers/Authentication/AdalAuthentication.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.IdentityModel.Clients.ActiveDirectory;
+using System;
 using System.Collections.Concurrent;
 using System.Management.Automation;
 using System.Threading.Tasks;
@@ -8,20 +9,25 @@ namespace PSSQT.Helpers.Authentication
 {
     class AdalAuthentication
     {
-        private static readonly ConcurrentDictionary<Guid, ADAL.TokenCache> Tokens = new ConcurrentDictionary<Guid, ADAL.TokenCache>();   // SPO Auth tokens
+        private static readonly ConcurrentDictionary<Guid, TokenCache> Tokens = new ConcurrentDictionary<Guid, TokenCache>();   // SPO Auth tokens
 
-        private static readonly string AuthorityUri = "https://login.windows.net/common/oauth2/authorize";
-        private static readonly string clientId = "9bc3ab49-b65d-410a-85ad-de819febfddc";
-        private static readonly string redirectUri = "https://oauth.spops.microsoft.com/";
+        protected static readonly string AuthorityUri = "https://login.windows.net/common/oauth2/authorize";
+        protected static readonly string clientId = "9bc3ab49-b65d-410a-85ad-de819febfddc";
+        protected static readonly string redirectUri = "https://oauth.spops.microsoft.com/";
 
-        private ADAL.AuthenticationContext AuthContext = null; //  new ADAL.AuthenticationContext(AuthorityUri);    // use default static cache - not thread safe?;
+        protected AuthenticationContext AuthContext = null; //  new ADAL.AuthenticationContext(AuthorityUri);    // use default static cache - not thread safe?;
 
         public AdalAuthentication()
         {
             CreateTokenCache();
         }
 
-        private void CreateTokenCache(bool forceRecreate = false)
+        //public static AdalAuthentication AdalAuthenticationFactory()
+        //{
+        //    return new AdalAuthentication();
+        //}
+
+        protected void CreateTokenCache(bool forceRecreate = false)
         {
             Guid runspaceId = Guid.Empty;
 
@@ -29,34 +35,40 @@ namespace PSSQT.Helpers.Authentication
             {
                 runspaceId = ps.Runspace.InstanceId;
 
-                ADAL.TokenCache tc;
 
-                bool found = Tokens.TryGetValue(runspaceId, out tc);
+                bool found = Tokens.TryGetValue(runspaceId, out TokenCache tc);
 
                 if (!found || forceRecreate)
                 {
-                    tc = new ADAL.TokenCache();
+                    tc = new TokenCache();
 
                     Tokens.AddOrUpdate(runspaceId, tc, (k, v) => v);     // Do I need to use a ConcurrentDictionary?
                 }
 
-                AuthContext = new ADAL.AuthenticationContext(AuthorityUri, tc);
+                AuthContext = new AuthenticationContext(AuthorityUri, tc);
             }
         }
 
-        public async Task<string> Login(string sharePointSiteUrl, bool forceLogin = false)
+        public virtual async Task<string> Login(string sharePointSiteUrl, bool forceLogin = false)
         {
             var spUri = new Uri(sharePointSiteUrl);
 
             string resourceUri = spUri.Scheme + "://" + spUri.Authority;
 
-            ADAL.AuthenticationResult authenticationResult;
+            AuthenticationResult authenticationResult = await AquireToken(resourceUri, forceLogin);
+
+            return authenticationResult.CreateAuthorizationHeader();
+        }
+
+        protected virtual async Task<AuthenticationResult> AquireToken(string resourceUri, bool forceLogin)
+        {
+            AuthenticationResult authenticationResult;
 
             if (forceLogin)
             {
                 CreateTokenCache(true);
 
-                var authParam = new ADAL.PlatformParameters(ADAL.PromptBehavior.Always);
+                var authParam = new PlatformParameters(PromptBehavior.Always);
                 authenticationResult = await AuthContext.AcquireTokenAsync(resourceUri, clientId, new Uri(redirectUri), authParam);
             }
             else
@@ -65,20 +77,20 @@ namespace PSSQT.Helpers.Authentication
                 {
                     authenticationResult = await AuthContext.AcquireTokenSilentAsync(resourceUri, clientId);
                 }
-                catch (ADAL.AdalSilentTokenAcquisitionException)
+                catch (AdalSilentTokenAcquisitionException)
                 {
 
                     try
                     {
                         // prevent flashing of login window when credentials are valid
-                        var authParam = new ADAL.PlatformParameters(ADAL.PromptBehavior.Never);
+                        var authParam = new PlatformParameters(PromptBehavior.Never);
                         authenticationResult = await AuthContext.AcquireTokenAsync(resourceUri, clientId, new Uri(redirectUri), authParam);
                     }
-                    catch (ADAL.AdalException /* e */)
+                    catch (AdalException /* e */)
                     {
                         //Console.WriteLine(e);
 
-                        var authParam = new ADAL.PlatformParameters(ADAL.PromptBehavior.Auto);
+                        var authParam = new PlatformParameters(PromptBehavior.Auto);
                         authenticationResult = await AuthContext.AcquireTokenAsync(resourceUri, clientId, new Uri(redirectUri), authParam);
 
                     }
@@ -86,8 +98,24 @@ namespace PSSQT.Helpers.Authentication
 
             }
 
-
-           return authenticationResult.CreateAuthorizationHeader();
+            return authenticationResult;
         }
+    }
+
+    class AdalUserCredentialAuthentication : AdalAuthentication
+    {
+        public AdalUserCredentialAuthentication(UserCredential credentials) :
+            base()
+        {
+            this.Credentials = credentials;
+        }
+
+        public UserCredential Credentials { get; }
+
+        protected override async Task<AuthenticationResult> AquireToken(string resourceUri, bool forceLogin)
+        {
+            return await AuthContext.AcquireTokenAsync(resourceUri, clientId, Credentials);
+        }
+
     }
 }
