+++ Solutions/SharePoint.Search.QueryTool/PSSQT/ResultProcessor/AllPropertiesInlineResultProcessor.cs
@@ -1,12 +1,8 @@
 ﻿using SearchQueryTool.Model;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace PSSQT
+namespace PSSQT.ResultProcessor
 {
     class AllPropertiesInlineResultProcessor : AllPropertiesResultProcessor
     {
