+++ Solutions/SharePoint.Search.QueryTool/PSSQT/ResultProcessor/IQueryResultProcessor.cs
@@ -1,0 +1,15 @@
+﻿using SearchQueryTool.Model;
+using System;
+
+namespace PSSQT.ResultProcessor
+{
+    public interface IQueryResultProcessor
+    {
+        void Configure();
+
+        void Process(SearchQueryResult searchQueryResult);
+
+        bool HandleException(Exception ex, SearchQueryRequest searchQueryRequest);
+
+    }
+}
