+++ Solutions/SharePoint.Search.QueryTool/PSSQT/SearchSPSuggestionsCmdlet.cs
@@ -11,8 +11,7 @@ using System.Collections.Generic;
 using PSSQT.Helpers;
 using PSSQT.Helpers.Authentication;
 using System.Threading;
-
-
+using PSSQT.ResultProcessor;
 
 /**
  * <ParameterSetName	P1	P2
@@ -170,6 +169,9 @@ namespace PSSQT
         {
             searchRequest.CopyFrom(preset.Request);
         }
+
+
+                
 
         protected override void ExecuteRequest(SearchSuggestionsRequest searchRequest)
         {
